+++ diesel/src/pg/expression/mod.rs
@@ -11,7 +11,7 @@ mod date_and_time;
 /// imported by [`expression::dsl`](../../expression/dsl/index.html) when
 /// compiled with the `feature = "postgres"` flag.
 pub mod dsl {
-    #[doc(inline)] pub use super::array_comparison::any;
+    #[doc(inline)] pub use super::array_comparison::{any, all};
 
     pub use super::extensions::*;
 }